pyproject.toml
[project]
name = "hide"
version = "0.2.0"
description = "Copyright data embedding via steganography"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.4.0",
]

[project.scripts]
hide = "hide.cli:main"

[tool.pytest.ini_options]
pythonpath = [
  "src/"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
